services/project_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
pydantic==2.9.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9